# docs/openapi/community-map-api.openapi.yaml
openapi: 3.1.0
info:
  title: Chengdu Tongzilin Community Map API
  version: 0.1.0
  description: |
    Community Map API for mobile, Mini Program, Admin, and Postman debugging.

    Current status:
    - Local Koa HTTP routes and shared contracts are implemented for the listed endpoints.
    - CloudBase dev URL is documented for deployment target only; `/api` route and live acceptance still require verification.
    - `x-mock-user-id` is a local/dev actor header and is not a production authentication scheme.
servers:
  - url: http://127.0.0.1:8787
    description: local Koa dev server
  - url: https://cloud1-d7gxdk8t43bd639c0.service.tcloudbase.com/api
    description: CloudBase dev HTTP access service, pending `/api` route verification
...
components:
  securitySchemes:
    MockActorHeader:
      type: apiKey
      in: header
      name: x-mock-user-id
      description: Local/dev mock actor id, for example `user_001` for admin flows.
...
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            [
              VALIDATION_ERROR,
              UNAUTHORIZED,
              FORBIDDEN,
              NOT_FOUND,
              INTERNAL_ERROR
            ]
        message:
          type: string
        details: {}
    ApiFailure:
      type: object
      required: [success, error, requestId]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
        requestId:
          type: string
    SuccessEnvelope:
      type: object
      required: [success, data, requestId]
      properties:
        success:
          type: boolean
          const: true
        data: {}
        requestId:
          type: string
...
    User:
      type: object
      required:
        [_id, nickname, avatar_url, preferred_language, role_flags, status]
      properties:
        _id:
          type: string
        openid:
          type: string
        unionid:
          type: string
        nickname:
          type: string
        avatar_url:
          type: string
          format: uri
        phone:
          type: string
        preferred_language:
          type: string
          enum: [zh, en]
        role_flags:
          type: array
          items:
            type: string
            examples: [community_admin, system_admin]
        status:
          type: string
    AuthSession:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
    Unauthorized:
      description: Actor cannot be resolved
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailure"
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Mock login
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                mock_user_id:
                  type: string
                  example: user_001
                preferred_language:
                  type: string
                  enum: [zh, en]
      responses:
        "200":
          description: Login session
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get current actor
      security:
        - MockActorHeader: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...